Add option to run slash in tmux

+++ slash/parallel/parallel_manager.py
@@ -12,6 +12,7 @@ from ..exceptions import INTERRUPTION_EXCEPTIONS
 from ..conf import config
 from ..ctx import context
 from .server import Server
+from ..utils.tmux_utils import create_new_window, kill_tmux_session
 
 _logger = logbook.Logger(__name__)
 log.set_log_color(_logger.name, logbook.NOTICE, 'blue')
@@ -23,8 +24,8 @@ class ParallelManager(object):
     def __init__(self, args):
         super(ParallelManager, self).__init__()
         self.server = None
-        self.args = [sys.executable, '-m', 'slash.frontend.main', 'run'] + args
-        self.workers_num = config.root.run.parallel
+        self.args = [sys.executable, '-m', 'slash.frontend.main', 'run', '--parent_session_id', context.session.id] + args
+        self.workers_num = config.root.parallel.workers_num
         self.workers = {}
         self.max_worker_id = 0
         self.server_thread = None
@@ -40,11 +41,17 @@ class ParallelManager(object):
     def start_worker(self):
         worker_id = str(self.max_worker_id)
         _logger.notice("Starting worker number {}".format(worker_id))
-        with open(os.devnull, 'w') as devnull:
-            proc = subprocess.Popen(self.args[:] + \
-                    ["--worker_id", worker_id, "--master_session_id", context.session.id], stdin=devnull, stdout=devnull, stderr=devnull)
-            self.workers[worker_id] = proc
-            self.max_worker_id += 1
+        new_args = self.args[:] + ["--worker_id", worker_id]
+        if config.root.run.tmux:
+            new_args.append(';$SHELL')
+            command = ' '.join(new_args)
+            self.workers[worker_id] = create_new_window("worker {}".format(worker_id), command)
+        else:
+            with open(os.devnull, 'w') as devnull:
+                proc = subprocess.Popen(new_args, stdin=devnull, stdout=devnull, stderr=devnull)
+                self.workers[worker_id] = proc
+        self.max_worker_id += 1
+
 
     def start_server_in_thread(self, collected):
         self.server = Server(collected)
@@ -53,7 +60,7 @@ class ParallelManager(object):
         self.server_thread.start()
 
     def stop_server(self):
-        client = xmlrpc_client.ServerProxy('http://{0}:{1}'.format(config.root.run.server_addr, config.root.run.server_port))
+        client = xmlrpc_client.ServerProxy('http://{0}:{1}'.format(config.root.parallel.server_addr, config.root.parallel.server_port))
         client.stop_server()
 
     def start_workers(self):
@@ -66,19 +73,26 @@ class ParallelManager(object):
                 for worker_id in workers_last_connection_time:
                     delta = (datetime.now() - workers_last_connection_time[worker_id]).seconds
                     if delta > COMMUNICATION_TIMEOUT_SECS:
-                        _logger.error("Worker {} is down, restarting".format(worker_id))
-                        if self.workers[worker_id].poll() is None:
-                            _logger.error("Killing worker {}".format(worker_id))
-                            self.workers[worker_id].kill()
+                        if not config.root.run.tmux:
+                            _logger.error("Worker {} is down, restarting".format(worker_id))
+                            if self.workers[worker_id].poll() is None:
+                                _logger.error("Killing worker {}".format(worker_id))
+                                self.workers[worker_id].kill()
+                        else:
+                            self.workers[worker_id].rename_window('stopped_client_{}'.format(worker_id))
                         self.server.report_client_failure(worker_id)
                         self.start_worker()
                 time.sleep(COMMUNICATION_TIMEOUT_SECS)
         except INTERRUPTION_EXCEPTIONS:
-            for worker in self.workers.values():
-                worker.kill()
-            self.stop_server()
-            raise
+            if config.root.run.tmux:
+                kill_tmux_session()
+            else:
+                for worker in self.workers.values():
+                    worker.kill()
+                self.stop_server()
+                raise
         finally:
-            for worker in self.workers.values():
-                worker.wait()
-            self.server_thread.join()
+            if not config.root.run.tmux:
+                for worker in self.workers.values():
+                    worker.wait()
+                self.server_thread.join()
